Add `fail-fast: false` to API Compatibility GitHub Actions
This PR adds `fail-fast: false` to API Compatibility GitHub Actions.
The main reason is to make sure if any job fails, the parallel jobs
within the same matrix of the workflow can continue. The API Compatibility
is to see how our plugin binaries match with different versions
and as such we want to see the whole compatibility match up results.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>