Use cat > service_account_creds.json << EOF on GitHub Actions Windows build
On GitHub Actions, we need to dump GCP_CREDS to a file. We used to
use printenv GCP_CREDS and this works on Windows (as GitHub
Actions setup the bash already).

However, some recent change (in order to get the Visual Studio shell)
caused the printenv not to work directly on Windows.

This PR use `cat > service_account_creds.json << EOF` so that this works again.

This PR also standarize the GCP setup on all jobs.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>