Apply visibility=hidden further
This PR updates third-party libraries as much as possible
by passing macros or patch visibility defines so that
most of the third-party libraries can be compiled with symbols
hidden. This will significally reduce the chance of symbol collision.

Note boost is the only one left and it still exposes lots of symbols.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>