setup.py: move tensorflow requirement to extras_require
The AMD ROCm project maintains a fork of the Tensorflow package
called 'tensorflow-rocm', a drop-in replacement for 'tensorflow'.
The fact that tensorflow-io states only 'tensorflow' as specific
dependency makes it hard to use 'tensorflow-rocm', since both
don't work whell in the same environment.

This change follows what's done for tensorflow-addons in
https://github.com/tensorflow/addons/commit/61ebf40b1cd97506e4e43bdcf30c20c43d1597cc,
namely:
- Move the 'tensorflow' requirement to extras_require.
- Add new extras_require for tensorflow-{rocm,cpu,gpu}

Fixes #1548