Store release candidate packages to dockerhub
While we try to automate the release process, one issue we are facing
is: where to store the release candidate python packages?
In GitHub Actions, the python packages are stored as artifact in our workflow.
However, GitHub workflow does not allow downloading the artifact through
API, it only allows downloading the artifact through UI.
For that we need to have a place to save the wheel files (before we automatically
release a version).

This PR utilize the dockerhub to save the wheel files:
1. We package wheel files and information (commit id, sha256 of packages, and version)
   into a minimal docker container image `tfsigio/candidate:{version}` and push to dockerhub.
2. The next step is to write a release workflow so that when a tag is triggered,
   the workflow will automatically pull the docker image `tfsigio/candidate:{version}`,
   and use the saved wheel files to push to PyPI.org (and create release from GitHub API).

Will have a follow up PR once this PR is merged (as the docker image will have to be available
to test).

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>